+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## $v
 
+- Adding support for Markdown in Text Widgets starting with WordPress v4.8, via the new `widget_text_content` filter.
 - Adding Hash CSS framework compatibility.
 
 ## v170408.19959